Change GetBooksUseCase to interface pattern
- This makes the GetBooksUseCase.kt more suitable for mocking
